[refactor] cardShapeTypes 파일 이름 수정
